fix: delete balance.ts file and optimize code for packages/common
