+++ packages/common/src/chains/chains.ts
@@ -105,3 +105,10 @@ export const getAllSupportedChains = (): Chain[] => {
 export const isSupportedChain = (chainId: unknown): chainId is ChainId => {
   return typeof chainId === "string" && SUPPORTED_CHAINS.some(chain => chain.id === chainId)
 }
+
+export const CHAIN_PROPERTIES: Record<KnowChainId, { decimals: number; symbol: string }> = {
+  polkadot: { decimals: 10, symbol: 'DOT' },
+  west: { decimals: 12, symbol: 'WND' },
+  polkadot_asset_hub: { decimals: 10, symbol: 'DOT' },
+  west_asset_hub: { decimals: 12, symbol: 'WND' }
+}
