+++ packages/common/src/types/index.ts
@@ -2,4 +2,3 @@ export * from "./account"
 export * from "./smoldot"
 export * from "./xcm"
 export * from "./agent"
-export * from "./balance"
